Initial page refinement
Still to resolve split question on p79.
Adding a line to qu 2 and 'start' tag to qu 3 results in all content moving to the next page.
Adding just the line results in numbering restarting at 1 for qu 3 (which lands on the next page).
Keep together also forces all content to the next page.